Added ability to thread incomming bluetooth on PC
All PC bluetooth receive code should now be threaded and communicate
back to the main thread.  Still need to implement stopping the incomming
message thread. This code also needs to be tested to make sure that it
actually works.  Also for some reason the project for the UML
documentation was being built, so I removed it from the build settings
since it contains no code.